requirements.txt
beautifulsoup4==4.12.3 
docx==0.2.3
pandas==2.2.2
requests==2.32.3 
python-docx==1.1.2 
alive-progress==3.1.5 
tabulate2==1.9.0 
colorama==0.4.6 